# document-rag-platform/packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnswerClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_text": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Claim Text",
            "type": "string"
          },
          "source_labels": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Labels",
            "type": "array"
          }
        },
        "required": [
          "claim_text",
          "source_labels"
        ],
        "title": "AnswerClaim",
        "type": "object"
      },
...
      "ChatQuery": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "debug": {
            "default": false,
            "title": "Debug",
            "type": "boolean"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "scope": {
            "default": "all",
            "enum": [
              "all",
              "documents",
              "images",
              "code"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "query",
          "project_id"
        ],
        "title": "ChatQuery",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "answerable": {
            "title": "Answerable",
            "type": "boolean"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "title": "Citations",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/AnswerClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "no_answer_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NoAnswerReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "retrieval_debug": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Debug"
          },
          "safety_flags": {
            "items": {
              "type": "string"
            },
            "title": "Safety Flags",
            "type": "array"
          },
          "uncertainty": {
            "items": {
              "type": "string"
            },
            "title": "Uncertainty",
            "type": "array"
          }
        },
        "required": [
          "conversation_id",
          "answer",
          "answerable",
          "no_answer_reason",
          "claims",
          "uncertainty",
          "safety_flags",
          "citations",
          "retrieval_debug"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "CitationResponse": {
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Name"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "heading_path": {
            "items": {
              "type": "string"
            },
            "title": "Heading Path",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "line_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line End"
          },
          "line_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "source_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File Id"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "symbol_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol Name"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "label",
          "document_id",
          "document_name",
          "version_id",
          "source_file_id",
          "source_type",
          "heading_path",
          "page_start",
          "page_end",
          "file_path",
          "symbol_name",
          "line_start",
          "line_end",
          "bbox",
          "snippet",
          "rank"
        ],
        "title": "CitationResponse",
        "type": "object"
      },
      "DiagnosticRank": {
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "chunk_id",
          "rank",
          "score"
        ],
        "title": "DiagnosticRank",
        "type": "object"
      },
      "DiagnosticsResponse": {
        "properties": {
          "bundle_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Hash"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "retrieval_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Run Id"
          },
          "stages": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/DiagnosticRank"
              },
              "type": "array"
            },
            "title": "Stages",
            "type": "object"
          },
          "timings_ms": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Timings Ms",
            "type": "object"
          }
        },
        "required": [
          "retrieval_run_id",
          "bundle_hash",
          "stages",
          "fallback_reason",
          "timings_ms"
        ],
        "title": "DiagnosticsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoAnswerReason": {
        "enum": [
          "smalltalk",
          "policy_refusal",
          "insufficient_evidence",
          "provider_failure",
          "permission_denied",
          "malformed_response"
        ],
        "title": "NoAnswerReason",
        "type": "string"
      },
...
      "RetrievalDebugRequest": {
        "additionalProperties": false,
        "properties": {
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "project_id"
        ],
        "title": "RetrievalDebugRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Document RAG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat/query": {
      "post": {
        "operationId": "query_chat_api_v1_chat_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/v1/debug/retrieval": {
      "post": {
        "description": "Run retrieval diagnostics without exposing raw context content.",
        "operationId": "debug_retrieval_api_v1_debug_retrieval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalDebugRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Debug Retrieval",
        "tags": [
          "debug"
        ]
      }
    },
...
  }
}